% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{chr_to_numeric}
\alias{chr_to_numeric}
\title{Convert Chromosome Name to Numeric}
\usage{
chr_to_numeric(chrString)
}
\arguments{
\item{chrString}{Character string of convention TODO}
}
\value{
TODO
}
\description{
Converts character string of format e.g. chr10 into numeric 10, or chrX into
23
}
