% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{cnv_genomeplots_per_patient}
\alias{cnv_genomeplots_per_patient}
\title{Conumee CNV Genomeplots}
\usage{
cnv_genomeplots_per_patient(outputDir, samplesheet, detailRegions = NULL)
}
\arguments{
\item{outputDir}{Directory that contains all intratumormeth output.}

\item{samplesheet}{Relates sample id to patient id and more.}

\item{detailRegions}{Character vector containing names of those detail
regions that should appear as labels on the CNV plot.}
}
\description{
Render a PDF containing conumee CNV genomeplots for every sample organized by
patient.
}
