% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_betas}
\alias{get_betas}
\title{Get Methylation Beta Values}
\usage{
get_betas(
  outputDir,
  probeFilter = NULL,
  sampleFilter = NULL,
  normalization = "raw",
  removeNaProbes = FALSE
)
}
\arguments{
\item{outputDir}{Output directory of the intratumormeth study.}

\item{probeFilter}{Character vector containing probe ids that should be
filtered out.}

\item{sampleFilter}{Character vector containing sample ids that should be
filtered out.}

\item{normalization}{Should be either "raw", "noob", or "swan".}

\item{removeNaProbes}{If TRUE, beta values of probes that are NA for at least
one sample will be removed.}
}
\value{
Matrix containing methylation beta values with CpG probes as rows and
sample ids as columns.
}
\description{
Loads the previously generated methylation beta values from disk and retuns
them in a matrix. This function allows to specify probe and sample filters.
These should be specified as character vectors containing probe/sample ids.
Furthermore, the normalization method can be specified. Probes that contain
NA values can be removed from all samples by specifying \code{removeNaProbes = TRUE}.
}
