% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{postprocess_conumee_files}
\alias{postprocess_conumee_files}
\title{Postprocess Conumee Files}
\usage{
postprocess_conumee_files(outputDir)
}
\arguments{
\item{outputDir}{Directory where output of intratumormeth is stored.}
}
\description{
Extract the relevant information from conumee CNV.analysis objects.
}
