% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{predict_sex}
\alias{predict_sex}
\title{Predict Sex Based on Methylation Data}
\usage{
predict_sex(outputDir, normalization = "raw", predictionCutoff = -2)
}
\arguments{
\item{outputDir}{TODO}

\item{normalization}{TODO}

\item{predictionCutoff}{What should the difference in log2 copynumber be
between males and females. See \code{?minfi::getSet} for details. Default
is -2.}
}
\value{
TODO
}
\description{
TODO describe whats happening here (difference between X and Y chr
intensities as indicator for male/female samples)
}
