% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_methylation_data}
\alias{preprocess_methylation_data}
\title{Preprocess Methylation Data}
\usage{
preprocess_methylation_data(samplesheet, outputDir, minfiForce = FALSE)
}
\arguments{
\item{samplesheet}{Dataframe containing columns 'sample_id' and 'Basename'.}

\item{outputDir}{Directory to save results to.}

\item{minfiForce}{Argument 'force' passed to minfi::read.metharray() call}
}
\description{
Under the hood this function mainly calls methods from the minfi package. It
applies different preprocessing methods and saves the resulting methylation
data to the disk under the specified directory.
}
