% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber.R
\name{run_conumee}
\alias{run_conumee}
\title{Run Conumee Workflow}
\usage{
run_conumee(outputDir, samplesheet, patientsheet, refF, refM, annoF, annoM)
}
\arguments{
\item{outputDir}{TODO}

\item{samplesheet}{TODO}

\item{patientsheet}{TODO}

\item{refF}{TODO}

\item{refM}{TODO}

\item{annoF}{TODO}

\item{annoM}{TODO}
}
\value{
TODO
}
\description{
Run Conumee Workflow
}
